remove chat page
